requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
asyncpg==0.30.0
httpx==0.27.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.9
PyJWT==2.8.0
bcrypt>=4.0.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.3
thefuzz==0.22.1
python-Levenshtein==0.26.1
langdetect==1.0.9
spacy==3.8.11
faiss-cpu==1.13.2